Backported YUI3's Y.log method.
YAHOO_config is used to enable/disable logging and if the browser console
is used, with both defaulting to true. If logging solely to the browser
console, the Logger widget is no longer required. If the console supports
logging levels such as "warn" and "error" than these are used if they
match the log category. Logging entries can be filtered by their src
using the YAHOO_config properties logInclude, for whitelisting, and
logExclude for blacklisting.